cambios en alertas y alinear logos

+++ src/main/resources/static/web/js/index.js
@@ -80,11 +80,31 @@ createApp({
     añadirCarrito(id){
       this.filtro_planta_carrito = this.plantas.filter(planta => planta.id == id)[0];  
       if (!(this.carrito.some(planta => planta.id == id))) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Producto añadido',
+          text: 'Se ha agregado a su carrito de comptras',
+          background:' rgb(238 243 236)',
+          confirmButtonColor: "#324545",
+          iconColor:"#324545",
+          
+        })
+          console.log("si lo esta añadiendo") 
           this.carrito.push(this.filtro_planta_carrito);
           this.totalCompra = this.carrito.reduce((acumulador, prod)=> acumulador += (prod.precio * prod.contador), 0)
           localStorage.setItem("carrito", JSON.stringify(this.carrito));
           localStorage.setItem("totalCompra", JSON.stringify(this.totalCompra))
-      }                
+      } else{
+        Swal.fire({
+          icon: 'info',
+          title: 'Usted ya añadio este producto al carrito de compras',
+          text: 'Para añadir más unidades diríjase al carrito de compras!',
+          background:' rgb(238 243 236)',
+          confirmButtonColor: " #324545",
+          iconColor:"#324545",
+          
+        })
+      }               
   },
     sumar(id) {
       this.carrito.map(planta => {
@@ -148,15 +168,30 @@ createApp({
         .catch(error => console.log(error))
     },
     logout() {
-      axios.post('/api/logout')
-        .then(response => {
-          this.carrito = [];
-          this.totalCompra = this.carrito.reduce((acumulador, prod) => acumulador += (prod.precio * prod.contador), 0)
-          localStorage.setItem("carrito", JSON.stringify(this.carrito));
-          localStorage.setItem("totalCompra", JSON.stringify(this.totalCompra))
-          window.location.href = '/index.html'
-        })
-        .catch(error => console.log(error))
+      Swal.fire({
+        title: 'Esta seguro de cerrar sesión?',
+        text: "Confirmar",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#324545',
+        cancelButtonColor: '#db3939',
+        background:' rgb(238 243 236)',
+        iconColor:"#324545",
+        confirmButtonText: 'Si, cerrar sesión!',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.post('/api/logout')
+          .then(response => {
+            this.carrito = [];
+            this.totalCompra = this.carrito.reduce((acumulador, prod) => acumulador += (prod.precio * prod.contador), 0)
+            localStorage.setItem("carrito", JSON.stringify(this.carrito));
+            localStorage.setItem("totalCompra", JSON.stringify(this.totalCompra))
+            window.location.href = '/index.html'
+          })
+          .catch(error => console.log(error))
+        }
+      })
     },
   }
 }).mount("#app")
